requirements.txt
asyncpg-lite~=0.22.2.1
aiogram~=3.7.0
python-decouple~=3.8